freya 0.4.0-rc.19

+++ src/views/file_explorer/file_explorer_ui.rs
@@ -313,7 +313,7 @@ impl Component for FileExplorer {
                         (items, focused_item_index, radio_app_state),
                         move |a, b| file_explorer_item_builder(a, channel.clone(), b),
                     )
-                    .length(length as i32)
+                    .length(length)
                     .item_size(27.)
                     .scroll_with_arrows(false),
                 )
